Organize G2P and D2P comparison modules
