+++ src/dismech/compare/__init__.py
@@ -1,0 +1,1 @@
+"""Comparison utilities and CLIs for external-vs-dismech audits."""
